afficher la section activite si c est seulement un mediateur
